refactor: remove pages from module entrypoints to avoid cyclic dependencies
