+++ src/features/events/index.ts
@@ -7,6 +7,3 @@ export { useCreateEvent } from './hooks/useCreateEvent'
 
 export * from './contexts/event-filter'
 export * from './contexts/event-view'
-
-export * from './pages/CreateEventPage'
-export * from './pages/DashboardPage'
